x86/amd: Configure necessary MSRs for SNP during CPU init when running as a guest
Hyper-V may expose the SEV/SEV-SNP CPU features to the guest, but it is up to
the guest to use them. early_detect_mem_encrypt() checks SYSCFG[MEM_ENCRYPT]
and HWCR[SMMLOCK] and if these are not set the SEV-SNP features are cleared.
Check if we are running under a hypervisor and if so - update SYSCFG and skip
the HWCR check.

It would be great to make this check more specific (checking for Hyper-V) but
this code runs before hypervisor detection on the boot cpu.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@linux.microsoft.com>