crypto: ccp: support probing vpsp device on x86
vpsp is a platform device so we need to hook up the registration of the
platform_driver portion. vpsp does not use static register offsets but
instead requires parsing the ASPT ACPI table for offests. Right now
interrupts are not supported because I have not figured out how to
allocate a vector for it.