crypto: ccp: always reclaim pages when running on vpsp
The RMP table is used to manipulate whether pages belong to hypervisor,
firmware or guest (and lots of other attributes). On baremetal, the PSP will
manipulate some rmp attributes on some commands, so no reclaim is required, but
the vpsp emulation seems to always expect a reclaim command to be sent before
rmpupdating pages to belong to host.