arch/x86/kernel/sev: overprovision shadow rmptable size
The shadow rmptable has an array length of totalram_pages(), but is
indexed using pfn. It ends up being too small, probably because the
memory map is sparse and pfns may be greater than totalram_pages(). It
is likely that this will require taking into account a max_pfn() or
something equivalent to properly size the allocation.