arch/x86/kernel/sev: update all rmp entries when hugepage is rmpupdated
The previous logic was too simple and left entries in a bad state when a
hugepage is assigned or unassigned. We need to update the individual
pages assigned bit, and unmark the pagesize when a hugepage is assigned
back to the hypervisor. This logic comes from APM volume 3, rmpudpate
instructions.

If we don't mark individual page as assigned when the whole hugepage is
assigned, we won't correctly switch the page to shared when it's
private. If we don't unset pagesize when freeing a hugepage, then
we will hit errors on rmpupdate which leads to memory failure.