x86/psp: Register PSP platform device when ASP table is present
The ASP table contains the memory location of the register window for
communication with the Platform Security Processor. The device is not exposed
as an acpi node, so it is necessary to probe for the table and register a
platform_device to represent it in the kernel.

At least conceptually, the same PSP may be exposed on the PCIe bus as well, in
which case it would be necessary to choose whether to use a PCI BAR or the
register window defined in ASPT for communication. There is no advantage to
using the ACPI and there are no known bare-metal systems that expose the ASP
table, so device registration is restricted to the only systems known to
provide an ASPT: Hyper-V VMs. Hyper-V VMs also do not expose the PSP over PCIe.

This is a skeleton device at this point, as the ccp driver is not yet prepared
to correctly probe it. Interrupt configuration will come later on as well.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@linux.microsoft.com>