crypto: ccp - Allow platform device to be psp master device
Move the getters/setters to sp-dev.c, so that they can be accessed from
sp-pci.c and sp-platform.c. This makes it possible for the psp platform_device
to set the function pointers and be assigned the role of master device by
psp_dev_init().

While the case of a system having both a PCI and ACPI PSP is not supported (and
not known to occur in the wild), it makes sense to have a single static global
to assign to to keep the code sane. Should such a system occur, the logic in
psp_set_master() is that the pci device is preferred.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@linux.microsoft.com>