x86/hyperv: Allocate RMP table during boot
Hyper-V VMs can be capable of hosting SNP isolated nested VMs on AMD CPUs. One
of the pieces of SNP is the RMP (Reverse Map) table which tracks page
assignment to firmware, hypervisor or guest. On bare-metal this table is
allocated by UEFI, but on Hyper-V it is the respnsibility of the OS to allocate
one if necessary. The nested_feature 'HV_X64_NESTED_NO_RMP_TABLE' will be set
to communicate that no rmp is available. The actual RMP table is exclusively
controlled by the Hyper-V hypervisor and is not virtualized to the VM. The SNP
code in the kernel uses the RMP table for its own tracking and so it is
necessary for init code to allocate one.

While not strictly necessary, follow the requirements defined by "SEV Secure
Nested Paging Firmware ABI Specification" Rev 1.54, section 8.8.2 when
allocating the RMP:

- RMP_BASE and RMP_END must be set identically across all cores.
- RMP_BASE must be 1 MB aligned
- RMP_END – RMP_BASE + 1 must be a multiple of 1 MB
- RMP is large enough to protect itself

The allocation is done in the init_mem_mapping() hook, which is the earliest
hook I found that has both max_pfn and memblock initialized. At this point we
are still under the memblock_set_current_limit(ISA_END_ADDRESS) condition, but
explicitly passing the end to memblock_phys_alloc_range() allows us to allocate
past that value.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@linux.microsoft.com>