x86/sev: Add support for NestedVirtSnpMsr
The rmpupdate and psmash instructions, which are used in AMD's SEV-SNP to
update the RMP (Reverse Map) table, can't be trapped. For nested scenarios, AMD
defined MSR versions of these instructions which can be emulated by the
top-level hypervisor. One instance where these MSRs are used are Hyper-V VMs
which expose SNP isolation features to the guest.

Use the MSR versions of rmpupdate/psmash when the the NestedVirtSnpMsr CPU
feature is available. The MSRs are defined in "AMD64 Architecture Programmer’s
Manual, Volume 2: System Programming", section 15.36.19.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@linux.microsoft.com>