x86/sev: Maintain shadow rmptable on Hyper-V
Hyper-V can expose the SEV-SNP feature to guests, and manages the system-wide
RMP (Reverse Map) table. The SNP implementation in the kernel needs access to
the rmptable for tracking pages and deciding when to issue rmpupdate/psmash.
When running as a Hyper-V guest with SNP support, an rmptable is allocated by
the kernel during boot for this purpose. We need to keep this table in sync
with rmpupdates. Update the rmptable on successful rmpupdate/psmash insruction.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@linux.microsoft.com>