x86/sev: Maintain shadow rmptable on Hyper-V
Hyper-V can expose the SEV-SNP feature to guests, and manages the system-wide
RMP (Reverse Map) table. The SNP implementation in the kernel needs access to
the rmptable for tracking pages and deciding when/how to issue
rmpupdate/psmash.  When running as a Hyper-V guest with SNP support, an
rmptable is allocated by the kernel during boot for this purpose. Keep the
table in sync with issued rmpupdate/psmash instructions.

The logic for how to update the rmptable comes from "AMD64 Architecture
Programmer’s Manual, Volume 3" which describes the psmash and rmpupdate
instructions. To ensure correctness of the SNP host code, the most important
fields are "assigned" and "page size".

Signed-off-by: Jeremi Piotrowski <jpiotrowski@linux.microsoft.com>