crypto: ccp - Introduce quirk to always reclaim pages after SEV-legacy commands
On Hyper-V, the rmp_mark_pages_shared() call after a SEV_PLATFORM_STATUS fails
with return code 2 (FAIL_PERMISSION) due to the page having the immutable bit
set in the RMP (SNP has been initialized). The comment above this spot mentions
that firmware automatically clears the immutable bit, but I can't find any
mention of this behavior in the SNP Firmware ABI Spec.

Introduce a quirk to always attempt the page reclaim and set it for the
platform PSP. It would be possible to make this behavior unconditional as the
firmware spec defines that page reclaim results in success if the page does not
have the immutable bit set.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@linux.microsoft.com>