KVM: x86: Add gmem hook for invalidating private memory
In some cases, like with SEV-SNP, guest memory needs to be updated in a
platform-specific manner before it can be safely freed back to the host.
Wire up these hooks to the .free_folio kvm_gmem_aops callback to allow
for special handling of this sort when freeing memory in response to
FALLOC_FL_PUNCH_HOLE operations and when releasing the inode, and go
ahead and define an arch-hook for x86 since it will be needed for
handling memory used for SEV-SNP guests.

Signed-off-by: Michael Roth <michael.roth@amd.com>