x86/amd: Skip early_rmptable_check when running on Hyper-V
Check if the kernel is running on Hyper-V and exit successfully from
early_rmptable_check to prevent the SNP feature from being disabled.
This happens on CPUs other than the BSP, because the early cpu init code
runs before cpuhp notifiers. On Hyper-V, the kernel allocates the
rmptable and cpuhp notifiers are used to propagate the rmp msr values.
The BSP already exits early because max_pfn is not available yet.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>