TMP: crypto: ccp - Add quirk to ignore SEV init
Hyper-V's PSP only supports a subset of commands required to run SNP
guests. SEV_INIT fails as it is not essential in running nested SNP
guests. Skip the rest of the init after successful SNP init to get rid
of an error print from the console and to allow kvm to bring up SNP
guests.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>